finished organizing products module
