+++ server/src/modules/product/product.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@nestjs/common";
-import { Category, Company, Product } from "@prisma/client";
+import { Category, Company, Product, ProductImage } from "@prisma/client";
 import { v2 as cloudinary } from "cloudinary";
 import { PrismaService } from "../prisma/prisma.service";
 import { CreateProductDto } from "./dto/create-propduct.dto";
@@ -10,10 +10,11 @@ import { ProductsByIdsDto } from "./dto/products-by-ids.dto";
 import { RelatedProductsDto } from "./dto/related-products.dto";
 import { SearchDataDto } from "./dto/search-data.dto";
 import { UpdateProductDto } from "./dto/update-product.dto";
-import { updateProduct } from "./queries/update-product";
 import {
+    deleteOldImagesQuery,
     featuredProductsQuery,
     filteredProductsQuery,
+    getOldImagesQuery,
     getProductByIdQuery,
     getProducts,
     getProductsByIdsQuery,
@@ -23,7 +24,7 @@ import {
     searchBarDataQuery,
     updateProductCategoryQuery,
 } from "./product.queries";
-import { subCategoriesQuery } from "./queries/utils";
+import { subCategoriesQuery } from "./utils/sql";
 
 @Injectable()
 export class ProductService {
@@ -201,8 +202,56 @@ export class ProductService {
         await Promise.all([createProduct, updateCategory]);
         return true;
     }
-    updateProduct(id: number, input: UpdateProductDto) {
-        return updateProduct(this.prisma, id, input);
+    async updateProduct(id: number, input: UpdateProductDto) {
+        const { variants, img_id, img_src, ...updateData } = input;
+        // if (!req.session.user?.role.includes(Role.EDITOR)) {
+        //     throw new AuthenticationError(
+        //         "You don't have permissions for this action",
+        //     );
+        // }
+        // if (input.name.length < 3) throw new UserInputError("Name is too short");
+
+        //update product, create product images and connection to variants
+        const newVariants = variants?.map((p_id) => {
+            return { id: p_id };
+        });
+        const newProductImages = img_id.map((img, i) => {
+            return { img_id: img, img_src: img_src[i] };
+        });
+
+        const productUpdate = this.prisma.product.update({
+            where: { id: id },
+            data: {
+                ...updateData,
+                images: {
+                    create: newProductImages,
+                },
+                variants: {
+                    connect: newVariants,
+                },
+                variantsRelation: {
+                    connect: newVariants,
+                },
+            },
+        });
+
+        //update relationbetween company and category
+        const categoryUpdate = this.prisma
+            .$queryRaw`${updateProductCategoryQuery(input)}`;
+
+        //delete old images
+        if (img_id.length) {
+            const oldImages = await this.prisma.$queryRaw<
+                ProductImage[]
+            >`${getOldImagesQuery(id)}`;
+            this.prisma.$queryRaw`${deleteOldImagesQuery(id)}`;
+            cloudinary.api.delete_resources(oldImages.map((i) => i.img_id));
+        }
+
+        const promiseArray = [productUpdate, categoryUpdate];
+        await Promise.all(promiseArray);
+
+        return true;
     }
     async deleteproduct(input: number) {
         // if (!req.session.user?.role.includes(Role.EDITOR)) {
