+++ server/src/modules/product/utils/sql.ts
